Separate IntervalCard into smaller components
