+++ packages/ui/src/IntervalCard/index.ts
@@ -1,5 +1,5 @@
 import IntervalCard from "./IntervalCard";
 
-export { IntervalCardState, IntervalCardVariant } from "./IntervalCard";
+export { IntervalCardState, IntervalCardVariant } from "./types";
 
 export default IntervalCard;
